Timing, worse case, double bars, fixed window width
Program now displays how long it took, in seconds, to sort the bars.
Also implemented a button to restart the program with a worst-case
scenario (i.e. bars sorted from largest to smallest), a button to double
the number of bars, and fixed window width based on number of bars and
spacing between the bars